Fix `ld`'s response file support for special files (part 2)
This is a follow-up to:

https://github.com/tpoechtrager/cctools-port/pull/131

… which introduced a bug: the `st_size` field returned by
does not necessarily accurately represent the length if the
file is a special file.  For example, on macOS the `st_size`
field appears to empirically be no larger than 64 KB for
special files, even if the actual input is larger than 64
KB.  Consequently, `ld` incorrectly truncates the input
arguments to 64 KB on macOS for large response files, which
defeats the purpose of response files (since they're
typically used to support large command lines).

For these special files, we don't have a good way to
ascertain the length of the input other than to `read` from
the input and see how many bytes we receive.  This means
that we can't allocate all the necessary memory we require
up front and instead we need to dynamically resize the
argument array as we read.