Fix `ld`'s response file support for special files (part 2)
This is a follow-up to:

https://github.com/tpoechtrager/cctools-port/pull/131

… which introduced a bug: the `st_size` field does not necessarily
accurately represent the length if the file is a special file.  For
example, on macOS the `st_size` field appears to empirically be no
larger than 64 KB for special files (presumably the size of some
buffer), even if the actual input is larger than 64 KB.  As a result,
`ld` was truncating the input arguments to 64 KB for large response
files, which defeats the purpose of response files (since they're
typically used to support large command lines).

More generally, this issue isn't specific to `fstat`, but rather
appears to be an issue with anything that uses a file descriptor
returned by `open`.  For example, `read` misbehaves in the exact same
way and refuses to read more than 64 KB from a special file that was
opened by `open` even if you try to repeatedly `read` from the file to
completion.

For these special files, we don't have a good way to ascertain the
length of the input because `fstat` won't work (that only works on
file descriptors returned by `open`, which misbehave on special files),
nor can we seek to the end to determine the length (because special
files might not support rewinding the input) so the first part of this
fix is to simply read from the input to the end and see how many bytes
receive.  This means that we can't allocate all the necessary memory we
require up front and instead we need to dynamically resize the argument
array as we read.

The second part of this solution is to use `fopen` / `fread` / `close`
on the unhappy path when `mmap` fails instead of using `open` / `read`
/ `close` since the latter operations misbehave on special files.